route for stripe is working without errors, refactored store container, shopping cart and checkout form
