+++ react-furniture-store/src/Components/CheckoutForm/CheckoutForm.jsx
@@ -2,32 +2,31 @@ import React, {Component} from 'react';
 import {CardElement, injectStripe} from 'react-stripe-elements';
 
 class CheckoutForm extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {complete: false};
-    this.submit = this.submit.bind(this);
-  }
+  // constructor(props) {
+  //   super(props);
+  //   this.submit = this.submit.bind(this);
+  // }
 
-  async submit(ev) {
-    let {token} = await this.props.stripe.createToken({name: "Name"});
-    let response = await fetch("/charge", {
-      //the fetch request will go to our server...
-      method: "POST",
-      headers: {"Content-Type": "text/plain"},
-      body: token.id
-    });
+  // async submit(ev) {
+  //   let {token} = await this.props.stripe.createToken({name: "Name"});
+  //   let response = await fetch("http://localhost:9000/charge", {
+  //     method: "POST",
+  //     headers: {"Content-Type": "text/plain"},
+  //     body: token.id
+  //   });
   
-    if (response.ok) this.setState({complete: true});
-  }
+  //   if (response.ok) this.setState({purchaseComplete: true});
+  //   console.log()
+  // }
 
   render() {
-    if (this.state.complete) return <h1>Purchase Complete</h1>;
-
+    if (this.props.purchaseComplete) return <h1>Purchase Complete</h1>;
+  
     return (
       <div className="checkout">
         <p>Would you like to complete the purchase?</p>
         <CardElement />
-        <button onClick={this.submit}>Pay {this.props.totalCost}</button>
+        <button onClick={this.props.submit.bind(this)}>Pay : {this.props.totalCost}</button>
       </div>
     );
   }
